short

+++ kalle/boll/boll.c
@@ -1,4 +1,4 @@
-void kalleBoll( int count )
+void kalleBoll( short count )
 {
     // la la la la la
 }
